innovation_project/Generate3DModel/requirements.txt
numpy==1.24.1
numpy-stl==3.0.0
certifi==2022.12.7
opencv-python==4.8.0.76
Pillow==9.4.0
python-utils==3.7.0
